Fix some suppressed warnings (#2170)
* Fix all suppressed -Wunused-but-set-variable

* Fix all suppressed -Wdeprecated-literal-operator

* Fix all suppressed -Wmissing-field-initializers

* Fix all suppressed -Wmissing-designated-field-initializers

* Fix suppressed -Wdeprecated-declarations in our code

* Fix all suppressed -Wunused-result

* Fix all suppressed -Wunused-function

* Remove outdated -Wlogical-op-parentheses suppression

* Fix all suppressed -Wdeprecated-enum-float-conversion

* Fix all suppressed -Wunused-lambda-capture

* Fix all suppressed -Wfloat-conversion

Suppression in archive-manager.cpp was stale.

* Fix all suppressed -Wimplicit-int-float-conversion

* Fix all suppressed -Wdeprecated-this-capture

+++ overlay/overlay.cpp
@@ -741,8 +741,8 @@ void TrafficStats::add_packet(td::uint64 size, bool in) {
 }
 
 void TrafficStats::normalize(double elapsed) {
-  out_bytes = static_cast<td::uint64>(out_bytes / elapsed);
-  in_bytes = static_cast<td::uint64>(in_bytes / elapsed);
+  out_bytes = static_cast<td::uint64>(static_cast<double>(out_bytes) / elapsed);
+  in_bytes = static_cast<td::uint64>(static_cast<double>(in_bytes) / elapsed);
   out_packets = static_cast<td::uint32>(out_packets / elapsed);
   in_packets = static_cast<td::uint32>(in_packets / elapsed);
 }
